More cleanup in aztables
Moved constant definitions to constants.go.
Moved struct definitions to models.go.
Moved internal helper funcs next to their types.
