+++ sdk/data/aztables/sas_query_params.go
@@ -9,25 +9,6 @@ import (
 	"time"
 )
 
-// SASProtocol indicates the SAS protocol
-type SASProtocol string
-
-const (
-	// SASProtocolHTTPS can be specified for a SAS protocol
-	SASProtocolHTTPS SASProtocol = "https"
-
-	// SASProtocolHTTPSandHTTP can be specified for a SAS protocol
-	SASProtocolHTTPSandHTTP SASProtocol = "https,http"
-)
-
-// PossibleSASProtocolValues returns the possible values for the SASProtocol const type.
-func PossibleSASProtocolValues() []SASProtocol {
-	return []SASProtocol{
-		SASProtocolHTTPS,
-		SASProtocolHTTPSandHTTP,
-	}
-}
-
 // FormatTimesForSASSigning converts a time.Time to a snapshotTimeFormat string suitable for a
 // SASField's StartTime or ExpiryTime fields. Returns "" if value.IsZero().
 func FormatTimesForSASSigning(startTime, expiryTime time.Time) (string, string) {
